# go-client-generated-initial/model_user.go
/*
 * Web Server
 *
 * A simple web server for managing users
 *
 * API version: 1.0
 * Contact: dmanor@redhat.com
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type User struct {
	Id string `json:"id,omitempty"`
	Name string `json:"name,omitempty"`
	Email string `json:"email,omitempty"`
}