# go-server-generated-initial/go/model_new_user.go
/*
 * Web Server
 *
 * A simple web server for managing users
 *
 * API version: 1.0
 * Contact: dmanor@redhat.com
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type NewUser struct {

	Name string `json:"name,omitempty"`

	Email string `json:"email,omitempty"`
}